refactor: Operator 외부 클래스 분리, 일부 변수명 변경 등
